sensor: support replace _TIDx in sensor name
The PLDM-stake proposes that the sensor name should include "_TIDx" as
the subfix. Because the sensor name will be displayed to the user, the
"_TIDx" string should be replaced by more meaning string such as
terminus name.
This commit supports replacing the "_TIDx" string in sensor name by
the configured string in eid_to_name.json file. The file format will
be:
{
    "eids":
    [
        {
            "eid":20,
            "prefix": true,
            "string": "S0"
        }
    ]
}
Where:
    eid: the terminus EID.
    prefix: true if use prefix, false if use subfix
    string: replace string

Tested:
1. Add the configuration in eid_to_name.json.
2. "_TIDx" string should be replaced by the configured string

Signed-off-by: ThuBaNguyen <thu@os.amperecomputing.com>
Change-Id: Ide6efcd822c676979bf3f60efe81c43635879468