terminus: set dateTime in discovering terminus
The BMC have to update the date time to MPRo in discovery phase. This
date time will be used to create the TimeStamp in RAS events or PLDM
PDR repo info. BMC will get the date time in Elapes property of
xyz.openbmc_project.Time.EpochTime interface in
/xyz/openbmc_project/time/bmc path. This commit supports calling PLDM
command to SetDateTime in discovery phase.

Tested:
1. Power off the host. Then power on.
2. Check the time stamp in PLDM PDR Repo info.

Signed-off-by: ThuBaNguyen <thu@os.amperecomputing.com>
Change-Id: Ifeafe9376117010d966bdd3ce7c718d63191d801