sensor: poll and update the compact numeric sensors
This commit supports poll mechanism to update the terminus PLDM
sensors. The compact numeric is covered in this commits.

The sensor timer for each terminus will poll the terminus sensors
after each configured interval. This interval will be configured use
poll-sensor-timer-interval package config. The interval is
milliseconds. The default value 1000 milliseconds will be applied.

Because the performance and the responding time for GetSensorReading
of terminus can be different. The sleep interval between
GetSensorReading commands will be added. This value can be configured
thru sleep-between-get-sensor-reading option. The option is
milliseconds. The default value is 10 milliseconds.

Tested:
1. Turn on the terminus. Its' sensors should be updated.
2. Stress the terminus to change the sensor values.
3. Confirm the changing of sensor values in D-Bus interfaces.

Signed-off-by: ThuBaNguyen <thu@os.amperecomputing.com>
Change-Id: I2e69f14cc1fa18d71c86f0e7e29a7ec13982d3e7