sensor: poll and update the numeric effecter sensors
This commit supports poll mechanism to update the numeric effecter
sensors. Because the numeric effecter is not changed without the user
actions. And the user can only change those effecters by setting the
value property of sensor D-Bus. So the effecter sensors will only be
updated one time after each power on of the terminus.

Tested:
1. Turn on the terminus. The terminus sensors should be updated.
2. Stress the terminus to change the sensor values.
3. Confirm the changing of sensor values in D-Bus interfaces.

Signed-off-by: ThuBaNguyen <thu@os.amperecomputing.com>
Change-Id: Ib79e0858f2cbe0eb0e48a3040f0bf067a7e85be4