Remove the PLDM sensor D-Bus object in objectMapper
When the sensor D-Bus objects are removed in PLDM interface the
objects still are existing in ObjectManager in D-Bus interface. Adding
value/thresholds D-Bus properties with "emit_no_signals" option in
PLDM interface causes the sensor objects are not removed in
objectManager interface. Change the setting to "defer_signals" to fix
that.

Tested:
1. Check the removed sensor D-Bus objects in PLDM interface.
2. Those sensor D-Bus paths should be removed in objectManager
interface.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>