Fix failure in setting effecter
Sometime, The value of Value D-Bus property of effecter is updated
before the propertyChanged signal in Value property is triggered this
cause calls PLDM command to settting effecter is bypassed. The cause
of this bypass is the setting value is the same as value in D-Bus
property.
Correct code to store the latest setting value of effecter. This value
will be used to compare with the new value instead of value in the
D-Bus interface.

Tested:
1. Set value of effecter use "ipmitool raw 0x3c 0x11 <>"
2. Use "pldmtool raw -m 0x14 -d 0x80 0x2 0x32 0x9B 0" to confirm about
the setting value.
3. The set value of ipmitool should effect on the responsed value of
the pldmtool command.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>