ampere: Remove BE to LE workaround to converting pldm numericSensorState event data
Before rebasing, PLDM numericSensorState event data is in BE byte
order. But the data type in PLDM code is in LE. That why there is some
workaround code in ampere boot-progress and PCIe hot plug event to do
the converting from BE to LE.

After rebasing the event data is already in LE order. Those workaround
code should be removed.

Tested:
1. Power off the host.
2. Clear the Redfish log.
3. Power on the host.
4. Check the Redfish log to see the boot progress.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>