pldmd: Migrate instance ID allocation to pldm::InstanceIdDb
This removes the use of `pldm::dbus_api::Requester` from around the
code-base. This makes progress towards removing the DBus API entirely
once all its consumers are converted to the libpldm instance ID APIs.

There was never a good reason for the code using the class to have
the knowledge that it was related to DBus anyway, so this is, in effect,
a double clean up improving separation of concerns.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>
Change-Id: I2d9397cae1b3c8c251c32e36ca520aad9c9b8cf6