GetPDRs: Correct nextRecordHandle value
Mpro responses different values of NextRecordHandle for multi-parts
PDRs with and without multi-parts PDRs support.
This commit provides a workaround for the BMC to both Mpro firmware,
which does not use the response value of NextRecordHandle from
multi-parts PDRs to calculate the next PDR handle, but based on the
current record handle.

Tested:
1. Verify the existing of all host sensors with the host firmware
images which do or do not support multi-parts PDRs.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>