bert_capture: prevent crashes
During BERT capture, the Mpro stopped the PLDM service after 1.2 seconds.
The GetSensorReading command will fail and be added to the retry process.
After BMC rebooted the host, the old EID instances will be removed and
make the sensorObj in processSensorReading() null. But there are still some
GetSensorReading commands in the retry process, then call
processSensorReading() and access the sensorObj null pointer.

So we should check the sensorObj before accessing this pointer to
prevent PLDM crashes.

Change-Id: Ie0566acee615490222c1b4daae0adbe218599a3c
Signed-off-by: Dung Cao <dung@os.amperecomputing.com>